v3 cow estimator
Use extent writer as abstraction layer for both estimation and ota
installation. Going over install operations will match the logic used
in delta_performer to accurattely size the cow. Extent_writer
abstraction takes care of chunking the data for us.

To account for FEC data and hash tree writes, we want to add a little
bit of overhead as AddRawBlocks calls are per verity hash tree level. In
the case of small partitions, we want to set the overhead to a hard
number rather than a percentile

To ensure that our estimator is backwards compatible for downgrade
OTA's, we should add labels after each operation is processed (although
this isn't needed for sizing a v3 cow).

Bug: 322279333
Test: OTA on cuttlefish

Change-Id: I46f9da30641778095ee8543d7627320e309b9359